+++ packages/bpk-component-banner-alert/index.js
@@ -1,0 +1,4 @@
+import BpkBannerAlert, { ALERT_TYPES } from './src/BpkBannerAlert'
+
+export { ALERT_TYPES }
+export default BpkBannerAlert