+++ packages/bpk-component-code/src/BpkCode.js
@@ -8,7 +8,7 @@ BpkCode.propTypes = {
   children: PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
-  ])
+  ]).isRequired
 }
 
 export default BpkCode