+++ packages/bpk-component-code/src/BpkCodeBlock.js
@@ -12,7 +12,7 @@ BpkCodeBlock.propTypes = {
   children: PropTypes.oneOfType([
     PropTypes.arrayOf(PropTypes.node),
     PropTypes.node
-  ])
+  ]).isRequired
 }
 
 export default BpkCodeBlock